moved logs_event
